+++ src/barfi/__init__.py
@@ -1,2 +1,2 @@
 from barfi.st_barfi import st_barfi as st_barfi
-from barfi.st_flow import st_flow as st_flow
+from barfi.flow import st_flow as st_flow
