+++ src/barfi/flow/block/__init__.py
@@ -1,0 +1,1 @@
+from barfi.flow.block.base import Block as Block
