+++ src/barfi/flow/block/interface.py
