+++ src/barfi/flow/schema/__init__.py
@@ -1,0 +1,1 @@
+from barfi.flow.schema.manage import SchemaManager as SchemaManager
