--- src/barfi/st_flow/block/__init__.py
@@ -1,1 +1,0 @@
-from barfi.st_flow.block.base import Block as Block
