--- src/barfi/st_flow/schema/__init__.py
@@ -1,1 +1,0 @@
-from barfi.st_flow.schema.manage import SchemaManager as SchemaManager
