+++ tests/test_schema_flow.py
@@ -1,7 +1,7 @@
 import pytest
 import json
 from dataclasses import asdict
-from barfi.st_flow.flow.types import build_flow_schema_from_dict
+from barfi.flow.flow.types import build_flow_schema_from_dict
 
 
 def load_schema():
